requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
pyyaml
uritemplate
django-oauth-toolkit

